Divide code in two files
